+++ test/spec/cider/nrepl/middleware/stacktrace_spec_test.clj
@@ -13,8 +13,7 @@
    (try (eval form)
         (catch Exception e
           e))
-   pprint
-   {}))
+   pprint))
 
 (defn stack-frames
   [form]